[Ide] Move ProjectSystem handling into its own class
